+++ system/config/config.php
@@ -328,5 +328,6 @@ $GLOBALS['TL_CONFIG']['disableCron']          = false;
 $GLOBALS['TL_CONFIG']['coreOnlyMode']         = false;
 $GLOBALS['TL_CONFIG']['addLanguageToUrl']     = false;
 $GLOBALS['TL_CONFIG']['doNotRedirectEmpty']   = false;
+$GLOBALS['TL_CONFIG']['useAutoItem']          = false;
 
 ?>