+++ system/modules/backend/dca/tl_settings.php
@@ -46,7 +46,7 @@ $GLOBALS['TL_DCA']['tl_settings'] = array
 	'palettes' => array
 	(
 		'__selector__'                => array('useSMTP'),
-		'default'                     => '{title_legend},websiteTitle,adminEmail;{date_legend},dateFormat,timeFormat,datimFormat,timeZone;{global_legend:hide},websitePath,characterSet,customSections,disableCron,minifyMarkup,gzipScripts;{backend_legend},resultsPerPage,maxResultsPerPage,staticFiles,staticSystem,staticPlugins,doNotCollapse;{frontend_legend},urlSuffix,cacheMode,rewriteURL,disableAlias,addLanguageToUrl,doNotRedirectEmpty;{security_legend:hide},allowedTags,debugMode,coreOnlyMode,displayErrors,logErrors,disableRefererCheck,disableIpCheck;{files_legend:hide},allowedDownload,validImageTypes,editableFiles,templateFiles,maxImageWidth,jpgQuality,gdMaxImgWidth,gdMaxImgHeight;{uploads_legend:hide},uploadPath,uploadTypes,uploadFields,maxFileSize,imageWidth,imageHeight;{search_legend:hide},enableSearch,indexProtected;{smtp_legend:hide},useSMTP;{modules_legend},inactiveModules;{timeout_legend:hide},undoPeriod,versionPeriod,logPeriod,sessionTimeout,autologin,lockPeriod;{chmod_legend:hide},defaultUser,defaultGroup,defaultChmod;{update_legend:hide},liveUpdateBase'
+		'default'                     => '{title_legend},websiteTitle,adminEmail;{date_legend},dateFormat,timeFormat,datimFormat,timeZone;{global_legend:hide},websitePath,characterSet,customSections,disableCron,minifyMarkup,gzipScripts;{backend_legend},resultsPerPage,maxResultsPerPage,staticFiles,staticSystem,staticPlugins,doNotCollapse;{frontend_legend},urlSuffix,rewriteURL,addLanguageToUrl,doNotRedirectEmpty,useAutoItem,disableAlias;{cache_legend:hide},cacheMode;{security_legend:hide},allowedTags,debugMode,coreOnlyMode,displayErrors,logErrors,disableRefererCheck,disableIpCheck;{files_legend:hide},allowedDownload,validImageTypes,editableFiles,templateFiles,maxImageWidth,jpgQuality,gdMaxImgWidth,gdMaxImgHeight;{uploads_legend:hide},uploadPath,uploadTypes,uploadFields,maxFileSize,imageWidth,imageHeight;{search_legend:hide},enableSearch,indexProtected;{smtp_legend:hide},useSMTP;{modules_legend},inactiveModules;{timeout_legend:hide},undoPeriod,versionPeriod,logPeriod,sessionTimeout,autologin,lockPeriod;{chmod_legend:hide},defaultUser,defaultGroup,defaultChmod;{update_legend:hide},liveUpdateBase'
 	),
 
 	// Subpalettes
@@ -190,25 +190,11 @@ $GLOBALS['TL_DCA']['tl_settings'] = array
 			'inputType'               => 'text',
 			'eval'                    => array('nospace'=>'true', 'tl_class'=>'w50')
 		),
-		'cacheMode' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['cacheMode'],
-			'inputType'               => 'select',
-			'options'                 => array('both', 'server', 'browser', 'none'),
-			'reference'               => &$GLOBALS['TL_LANG']['tl_settings'],
-			'eval'                    => array('tl_class'=>'w50')
-		),
 		'rewriteURL' => array
 		(
 			'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['rewriteURL'],
 			'inputType'               => 'checkbox',
-			'eval'                    => array('tl_class'=>'w50')
-		),
-		'disableAlias' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['disableAlias'],
-			'inputType'               => 'checkbox',
-			'eval'                    => array('tl_class'=>'w50')
+			'eval'                    => array('tl_class'=>'w50 m12')
 		),
 		'addLanguageToUrl' => array
 		(
@@ -221,6 +207,25 @@ $GLOBALS['TL_DCA']['tl_settings'] = array
 			'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['doNotRedirectEmpty'],
 			'inputType'               => 'checkbox',
 			'eval'                    => array('tl_class'=>'w50')
+		),
+		'useAutoItem' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['useAutoItem'],
+			'inputType'               => 'checkbox',
+			'eval'                    => array('tl_class'=>'w50')
+		),
+		'disableAlias' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['disableAlias'],
+			'inputType'               => 'checkbox',
+			'eval'                    => array('tl_class'=>'w50')
+		),
+		'cacheMode' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['cacheMode'],
+			'inputType'               => 'select',
+			'options'                 => array('both', 'server', 'browser', 'none'),
+			'reference'               => &$GLOBALS['TL_LANG']['tl_settings']
 		),
 		'allowedTags' => array
 		(